add the first commit of seal script
this script works in two modes, one is to "lock", which archives
and encrypts the current directory, and the other is to "unlock",
and decrypts and extracts the archive.

the script assumes that in the system gpg is installed and the
user has a gpg key pair. the second requirement is tar utility.

Changes to be committed:
	new file:   seal.sh