solve deletion of wallpaper bug
when the script is run multiple times in a second, some wallpapers
are deleted. In the current implementation, the script checks add
a random number to the wallpaper name to avoid overwriting, and it
is double protected by a while loop that create new uniq name
until it finds one that does not exist.

Changes to be committed:
	modified:   wallpaper-changer.sh